update lakeshores to apstools, keep old as bkp.
